refactor(rust): extract the auth0 providers as trait implementations
